Erjulix

Connect Erlang, Julia and Elixir

Documentation for Erjulix.

Erjulix.Erjulix

Erjulix.EvalServer

Erjulix.deserializek

Erjulix.eServer

Erjulix.genpasswd

Erjulix.getHostPort

Erjulix.getPort

Erjulix.pServer

Erjulix.serializek

Erjulix.Erjulix — Module

Erjulix

Julia module for communicating with Erlang and Elixir.

Erjulix.EvalServer — Method

EvalServer(port::Integer, mod::Module, key::AbstractString)

An EvalServer runs as a task with its own module namespace. It receives UDP message tuples from Erlang:

(:eval, term): where term is a Symbol or a String,

(:call, term, args): with term as above and args a Vector{Any},

(:set, atoms, vals): where atoms is a Symbol or a vector of them and vals is a value Any or a vector of them.

It then evaluates the messages in its namespace as

strings to parse or symbols to evaluate,

functions to execute with arguments or

variables to create or to assign values to.

It sends a result tuple back to the Erlang client.

The server finishes if it gets an "exit" or :exit message.

If !isempty(key), the UDP packages get sha-256 encoded/decoded with that key as JSON Web tokens.

Erjulix.deserializek — Method

Deserialize a binary, sha-256 decoded with key if it is not empty.

Erjulix.eServer — Method

eServer(host::IPAddr, port::Integer, key::AbstractString)
eServer(port::Integer)

Create a module and spawn an EvalServer listening to an UDP port in its namespace and return the module.

Erjulix.genpasswd — Method

Return a Base64 encoded random passwort of length len.

Erjulix.getHostPort — Method

Return the Sockets.InetAddr of a socket sock.

Erjulix.getPort — Method

Return an available port number ≥ start.

Erjulix.pServer — Method

pServer(port::Integer)
pServer(host::IPAddr, port::Integer, key::AbstractString)

Start a server listening to an UDP port and starting parallel EvalServers if requested.

Erjulix.serializek — Method

Serialize data, sha-256 encoded with key if it is not empty.