Erjulix

Connect Erlang, Julia and Elixir

Documentation for Erjulix.

Erjulix.EvalServer

Erjulix.eServer

Erjulix.pServer

Erjulix.EvalServer — Method

EvalServer(port::Integer, mod::Module)

An EvalServer runs as a task with its own module namespace. It receives UDP message tuples from Erlang:

(:eval, term): where term is a Symbol or a String,

(:call, term, args): with term as above and args a Vector{Any},

(:set, atoms, vals): where atoms is a Symbol or a vector of them and vals is a value Any or a vector of them.

It then evaluates the messages in its namespace as

strings to parse or symbols to evaluate,

functions to execute with arguments or

variables to create or to assign values to.

It sends a result tuple back to the Erlang client.

The server finishes if it gets an "exit" or :exit message.

Erjulix.eServer — Method

eServer(port::Integer)

Create a module with an EvalServer listening to an UDP port in its namespace and return the module.

Erjulix.pServer — Method

pServer(port::Integer)
pServer(host::IPAddr, port::Integer)

Start a server listening to an UDP port and starting parallel EvalServers if requested.